Connect new EME APIs to internal implementation.
The MediaKeys/Decryptor creation code is moved from ProxyDecryptor to
ContentDecryptorModuleFactory so it can be shared with both versions of the API.

There must be a translation between session objects in the EME APIs and
the session_id used internally. This requires more work; for now, only
one session is supported.

BUG=250048

Review URL: https://chromiumcodereview.appspot.com/17320011