Native Android accessibility.
This is the final changelist that completes the initial implementation.
When accessibility is on and script injection is turned off or
unavailable, constructs a BrowserAccessibilityManager for the
ContentViewCore that provides a native accessibility implementation.

BUG=242953
R=benm@chromium.org, bulach@chromium.org, dtrainor@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=207875

Review URL: https://chromiumcodereview.appspot.com/15741009