Send IN_DICTIONARY feedback when user has misspelling in custom dictionary
When a user types a word that is misspelled and is in the custom dictionary,
then the spelling service client will generate a spellcheck result for the
misspelling anyway. The feedback sender sends an IN_DICTIONARY message to the
spelling service to indicate that the user cannot see this spellcheck result
and, therefore, cannot provide any other feedback on the spellcheck suggestion
quality.

BUG=170514

Review URL: https://chromiumcodereview.appspot.com/15897005