Files.app: Fixed the scroll height of the scroll bar.
Previouslly, the upper limit of the button position was lower than the actual
upper limit.  This CL fixed this by updating the scaling factor from button
position to scroll position.

Along the way, this CL made both the margin of table view and grid view be set
in order to avoid mismatch the margin when switing the views.

BUG=253455
TEST=manually
R=yoshiki@chromium.org

Review URL: https://codereview.chromium.org/17593005