Fix the initialization queue in the Files.app's background page.
In case of refocusing a current window, the callback wasn't called, what caused freezing of the queue. The queue was waiting for the callback, which was never called. This prevented any further tasks added to the queue from being executed. This patch resolves this issue by adding the missing callback.

TEST=Insert a SD card twice. Try to play an mp3 file.
BUG=253680
R=yoshiki@chromium.org

Review URL: https://codereview.chromium.org/17625006