Remove MediaStreamDeviceThread in favor of using AudioThread.
Doesn't seem to be a good reason for this thread; remove it in favor
of using the thread provided by the AudioManager.

This also ensures we don't issue CoreAudio calls on a thread other
than the main thread for OSX.

Bonus: Making this change required converting several unittests to use
the new TestBrowserThreadBundle instead of manually creating threads.

BUG=158170
TEST=unit tests pass. device changes work.

Review URL: https://chromiumcodereview.appspot.com/17444005