[Telemetry] Fix flakiness in sunspider benchmark.
Previously we were only checking that the URL changed, and then assuming
the |output| variable would be available. That meant we could try to
access output during the page load, before it is available.

To fix this, we wait for output to be defined before reading it.

BUG=254073
NOTRY=True

Review URL: https://chromiumcodereview.appspot.com/17731002