Support accent keys
Mocks are here:
https://folio.googleplex.com/cros-vkey/#/07-vkey-longpress-E.png
accents keys should be summoned by long press on a key. If there is not accent keys associated with the long pressed key, nothing should happen.

To add accent keys, here are the steps(temporary, may change in the future):
1. create a template element in the corresponding keyset element (make sure the template is the first child of the keyset)
2. create a structure like this (take lower case letter e for example):
<kb-accent-set id="e" offset="2">
  <kb-accent-key>&#xUUUU</kb-accent-key>
  <kb-accent-key>&#xUUUU</kb-accent-key>
  <kb-accent-key class="active">e</kb-accent-key>
  <kb-accent-key>&#xUUUU</kb-accent-key>
</kb-accent-set>
where id is the same as the letter we long pressed; offset represent how many keys should display on the left side of the pressed key(in our case, two keys should display on the left side of e after long pressed). class active should apply to the key that overlay the long pressed key (in our case, the overlay key should be the same as the original key).
3. add accents attribute the the e kb-key element, e.g. <kb-key accents="e">e</kb-key>

BUG=247734

Review URL: https://chromiumcodereview.appspot.com/16781005