Made media::MediaLogEvent use base::TimeTicks instead of base::Time.
base::Time has the unfortumate ability to decrease (due to change of
timezone, user change, etc), a feature which makes it a poor choice
to use for logging.  base::TimeTicks does not have this disadvantage
so it replaces base::Time for logging purposes.

initial changes to time in media_internals

BUG=249212

Review URL: https://chromiumcodereview.appspot.com/17444015