This issue is only a move and updates to the includes. It's contingent on 17101030.
For security reasons, we make the utility process parse the Picasa database PMP files. A pre-requisite of this is moving most of the PMP parsing code to utility/. A few items need to remain in common so that the code in browser/ can use the IPC structs and constants.

BUG=151701
R=gbillock

Review URL: https://chromiumcodereview.appspot.com/17408006