Window focus fixes for Core Animation
When changing window focus various controls in the toolbar wouldn't
update. For example:
  - blue focus ring in the omnibox and find bar wouldn't hide
  - icons in buttons wouldn't change to the light gray version

The problem was that these controls weren't being invalidted. Fix
was to recursively invalidate all toolbar controls when switching
window focus.

BUG=245900
TEST=Run chrome with --use-core-animation. Open a new window and
verify that the blue focus ring in the omnibox of the old window
is hidden.

Review URL: https://chromiumcodereview.appspot.com/17645003