Add more tests to Chromium WebRTC testers.
These are tests that normally don't execute in the main waterfall
(except content_unittest).
The PyAuto tests are in the progress of being converted to browser_tests
and Telemetry tests, but we'd like to get them up here now so we can
retire our own machines running these tests.

Overview of the new tests added:
* PyAuto AppRTC: tests that the WebRTC reference application is
  working.
* PyAuto quality: performs video and audio quality analysis to catch
  regressions.
* content_unittests: subset of the main waterfall execution of this test.
  Used to catch regressions with recent WebRTC revisions in the
  chromium.webrtc.fyi waterfall.
* manual browser_tests and content_browsertests: Tests that normally
  don't run in the main waterfall, as they require audio/video devices.

Prerequisites:
The quality tests needs ffmpeg 0.11.1 and zxing
(https://code.google.com/p/zxing/) executables to be installed in the
PATH of the buildbot tester.
This will be done by:
* Win: Copying ffmpeg and zxing into build_internal\tools
* Mac: Installing ffmpeg into /usr/local/bin and copying zxing into
  ~/slavebin
* Linux: Installing ffmpeg into /usr/bin and copying zxing into
  ~/slavebin

TEST=local execution on Linux, Windows and Mac.
BUG=244550

Review URL: https://chromiumcodereview.appspot.com/16145002