Pepper: FileRef host support for external files.
This change adds support for external filesystems to the host side of FileRef.
This is quite a bit simpler than supporting internal filesystems, since many
operations aren't supported.

I restructured how permissions are checked to ensure that we'd use the same
permissions for internal and external filesystems (and fixed a bug where the
wrong path was being checked for create permissions in Rename in the process).

TBR=darin
BUG=225441

Review URL: https://chromiumcodereview.appspot.com/17385012