Presentation mode fixes for Core Animation
In presentation mode the tab strip wasn't visible. The problem was that
the tab strip was not layer backed so it couldn't overlap the content
view.

Fix was to make the tab strip layer backed. This CL also fixes a couple
of issues that cropped up with a layer backed tab strip:
  - controls in the tab strip wouldn't update when switching window focus
  - tabs wouldn't update their selected state was switching tabs

Fix for the first problem was to have TabStripView inherit from
BackgroundGradientView. The gradient view already listens to all the
correct notifications to update itself on window focus switches.

Fix for the second problem was to invalidate the tab on selection
changes.

BUG=245900
TEST=Run chrome with --use-core-animation. Enter presentation mode.
Verified that the tab strip was visible.
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/17640004