Simplify strings in gallery.
Before, there was a custom function which was converting class names, title, names to string identifiers by making the argumetn upper case and prepending the GALLERY_ prefix. As a result, it was hard to maintain, and impossible to reuse strings which do not have the GALLERY_ prefix.

TEST=Files.app's gallery and the video player should display correctly.
BUG=254317
R=hirono@chromium.org

Review URL: https://codereview.chromium.org/18014002