Moved the CloudPolicyClientRegistrationHelper to its own files.
This class contacts Gaia to determine if an account has management enabled, and
if so it registers for cloud policy with the policy service. It also mints
OAuth2 tokens if necessary for these requests.

It used to be used only by the UserPolicySigninService for the desktop
platforms. This CL moves it to its own files, so that it can be reused on other
platforms too.

The next platform to use this is Android, which mints its own tokens on the Java
side. The CloudPolicyClientRegistrationHelper has been extended with a new
StartRegistrationWithServicesToken() which receives an appropriate OAuth2 token
from the caller.

BUG=249375

Review URL: https://chromiumcodereview.appspot.com/18029003