gitdeps: Only fetch repo if needed
Relanding earlier attempt to reduce fetches.

This does not work perfectly -- about 8 repos have DEPS pinned
to a revision which does not exist in the subfolder -- in this
case we cannot tell if we are missing the revision because of
a lost fetch, so we have to fetch. Once we use the same repos
as git_updater we can elimnate the fetches.