Apply cross-origin auth prompt blocking to image resources only.
This CL applies both cross-origin basic auth prompt blocking, and
the stripping of embedded URL identities to image src URLs only.
See the bug for a description of why we think this is acceptable.

The first part of the change is to rename HttpAuthResourceType to
HttpAuthRelationType, to avoid confusion with ResourceType since
both are now used closely together.

Then, BuildLoadFlagsForRequest() is made a method, rather than just
a function, so it can access information needed to compute the relation
type.  It blocks the sending of credentials as well as the popping 
of popups for images under the right conditions.

BUG=174179

Review URL: https://chromiumcodereview.appspot.com/17738004