Add SkColorToSRGBNSColor to skia_mac_utils
When WCS is enabled, the sRGB color space is the default color space for
Windows [1] and for CSS3 rgb values [2]. Using
SkColorToCalibratedNSColor results in a different (typically ligher)
color when used in Quartz drawing versus a web page, or for color
constants from designers.

This adds SkColorToSRGBNSColor to convert an SkColor value specified in
the sRGB color space to an NSColor.

[1] http://msdn.microsoft.com/en-us/library/windows/hardware/gg487409.aspx
[2] http://www.w3.org/TR/css3-color/#rgb-color

BUG=138633,  254361

Review URL: https://chromiumcodereview.appspot.com/17834003