Remove unused ManagedState::is_observed_ property
This is some code cleanup in preparation of adding a FavoriteState.

We used to only observe watched services, however this caused problems
so the logic was changed, but the property was not removed.

A side effect of not removing the property is that RequestProperties
was getting called twice when a service is added to Services, once from
ShillPropertyHandler and once from NetworkStateHandler. The effect is
a no-op since RequestProperties checks for pending updates, but it makes
the code more confusing.

BUG=251922
R=gauravsh@chromium.org, pneubeck@chromium.org

Review URL: https://codereview.chromium.org/18031008