[NaCl SDK] Fixup GCC-generated .d files (second attempt)
The change adds a script that will post-procress
gcc-generated .d files such that they do not have
file-not-found issue when headers are removed or
renamed.

Also, fix an issue in common.mk where it was checking
for IGNORE_DEPS incorrectly.  This was causing dependencies
to always be built which was leading to duplicate builds
of the same project causing race conditions in the build
system.

R=binji@chromium.org
BUG=None

Review URL: https://codereview.chromium.org/17769006