[sync] Hide "Sync nothing" UI; Go back to signing out on dashboard clears and abandoned initial setup
A new "Sync nothing" option was recently added to the advanced sync
settings dialog. It allowed the user to stop syncing, but stay signed
in. In addition, sign in and sync were separated, with the result that
on a dashboard clear, the user would remain signed in, but would no
longer have sync enabled.

For M29, we've decided to hide this UI and revert to old sign-in/out
behavior because there are other services that depend on the sync
engine being up and running.

This patch undoes a small number of these changes as listed below:

1) Hides the "Sync nothing" item in the advanced sync settings drop-down
2) Signs out the user if the user clicks "Choose what to sync" during
   initial setup, and then clicks cancel on the advanced dialog
3) Signs out the user when a dashboard clear is done
4) Signs out the user if the user clicks cancel while re-setting up sync
   after the dashboard clear
5) Reverts the status string shown when sync is cleared via the dashboard

BUG=235633, 252049, 248877
TEST=No "Sync nothing" UI in advanced settings; user is signed out when sync is disabled on Desktop; Chrome OS strings back to M28

Review URL: https://chromiumcodereview.appspot.com/17392009