Fix stream textures + mailboxes
DestroyStreamTexture was called from DeleteTextures, which is awkward because it doesn't mean the texture is actually deleted.
Instead, rationalize StreamTexture ownership, by making it associated with the TextureRef (i.e. client id) that created it. When the TextureRef is destroyed and the stream texture wasn't explicitly destroyed yet, it gets destroyed implicitly at this point.
The StreamTextureManager still maps StreamTextures by service id, so that multiple TextureRef in the same ContextGroup (in practice, in the same channel actually) will still allow updates on bind.

This also makes explicit the fact that the StreamTextureManager is shared per ContxtGroup, since we rely on that for Update.

BUG=254494

Review URL: https://chromiumcodereview.appspot.com/18430002