drive: Remove ReadDirectoryWithSettingCallback
ReadDirectoryWithSettingCallback was a hack needed to pass the hide hosted
documents setting to FileSystemProxy, so hosted files can be filtered inside
FileSystemProxy.

We should stop handling the setting in C++, but as a first step, remove
this callback type and move the filtering logic from FileSystemProxy to
FileSystem.

BUG=255335
TEST=hosted documents can be hidden or shown as before per the setting in the gear menu on Drive in Files.app

TBR=hidehiko@chromium.org
# for a minor change in chrome/browser/ui/webui/feedback_ui.cc

Review URL: https://codereview.chromium.org/18471002