Translate: Disable the "Don't translate" checkbox when the default target language is selected.
Chrome should not translate webpages which are written in the 'default target language'. In most cases, the 'default target language' is the user's locale language. When the language is selected in the language settings, the "Don't translate" checkbox should be always checked and the user should not be able to configure it.

BUG=253794
TEST=browser_tests

Review URL: https://chromiumcodereview.appspot.com/17812002