cras_unified: Add default case to stream-type switch.
A new value will be added to this enum for whole-system loopback.
That case will be added here after it is added to the system enum.
Until then, add a default case so that the compiler doesn't complain.

This will allow the enum to be changed, after that the new value can
be handled in this switch.

BUG=247160
TEST=Builds agains cros platform with new enum value defined

Review URL: https://chromiumcodereview.appspot.com/18344011