Iterate over listeners when sending process died notification.
The refactor in http://crrev.com/207913 moved the RenderWidgetHost iterator outside of RenderProcessHost. Since it is now a vector of pointers, which is not updated when RenderWidgetHosts are removed, iterating through those causes us to access already freed objects.
We should be iterating over all listeners and notifying those. In addition, it will fix the case where we have non-RWH listeners, which also need to be cleaned up.

BUG=255524

Review URL: https://chromiumcodereview.appspot.com/18167004