[NaCl SDK] Simplify SDK build scripts.
Don't pass the name of the current platform to all
functions.  This is essentially a global that can be
accessed from anywhere.  No need to pass this or the
verbosity level in arguments.  This reduces that arity
of many functions and makes the code more readable.

Also, remove cases where GetPlatform() should have
been used in place of comparing with sys.platform.

Also, remove duplicate code for calling invoking make.
Construction of the make commandline is now done solely
in build_projects.py.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/18563003