Convert WebIDBDatabaseImpl to IndexedDBConnection
Simplify what was a proxy class with a little bit of state to just the state, and reveal its true nature as the object that binds a db callback to a database - i.e. the actual connection object. Track this, rather than the db callback, in the database itself so that the a database can operate on connections (e.g. force close them all in case of backing store error)

BUG=234278
R=dgrogan@chromium.org
TBR=jam@chromium.org

Review URL: https://chromiumcodereview.appspot.com/18221003