TextureLayer: don't expect a context for mailboxes.
In the non-mailbox path, we use the context for detecting if the context has
been lost, to avoid races on recreation (context mismatch). This is not needed
with mailboxes, and the originating context may not even be available (e.g.
in another other process). So don't expect a context in that path.

BUG=255665

Review URL: https://chromiumcodereview.appspot.com/18420015