Fix task execution order between URLFetcherCore::DisownFile and Stop.
URLFetcher cancellation is currently implemented as:

if(now on IO thread) CancelURLRequest(); else PostTask(&CancelURLRequest);

and due to that, the order of execution may be reversed with another
operation always PostTask'ing. This is particularly problematic when
Cancel gets ahead of DisownFile(). In that case, disowning never
happens (actually even worse, current implementation just crashes).

This patch changes DisownFile *not* to post, in order not to
rely on any task scheduling to ensure correctness.

BUG=256109

Review URL: https://chromiumcodereview.appspot.com/18033003