Support network Favorite/Preferred list and removal
This CL tracks Manager.ServiceCompleteList to provide a list of
favorite/preferred networks to the UI. It uses 
Service.GetLoadableProfileEntries to remove preferred services.

A test UI is provided in the status area (behind a flag).

Stub behavior is updated to enable testing the UI on Linux.

BUG=251922

Review URL: https://chromiumcodereview.appspot.com/17778003