Changed the semantics of RenderWidgetHost::GetRenderWidgetHosts() to filter out swapped-out ones.
Changed the semantics of RenderWidgetHost::GetRenderWidgetHosts() so
that it only returns *active* RenderWidgetHosts which are not swapped
out. We (creis/dsjang) checked all the callsites of this function.

This change fixes the bug that swapped-out views are shown (they are
not supposed to be shown) on 1) chrome://memory, 2)
chrome://performance (in an event of killing a view), 3)
chrome://accessibility.

BUG=137990

Review URL: https://chromiumcodereview.appspot.com/17941005