Provide support in DSC for Dictionary
Alows a LIST such as CCFLAGS or LIBS to be defined as a DICT.
When a LIST is passed in, the LIST applies to all keys of the
DICT.  This is implemented in template.mk.

Also puts logic for removing unused values in the same
expansion path.

R=sbc@chromium.org
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/18301009