[NaCl SDK] Fix for buiding html files in pnacl configurations.
I found this issue while using the common.mk system
to build the mesagl example in naclports:
https://codereview.chromium.org/18404005

I also changed the way that the $(PAGE) existence is
checked.  The error message is a little less informative
(no rule to make index.html) but it is produced more
consistently.  Mixing $(error) with make rules can lead
to confusing output since the "ifdef" check is done at
make parse time but the error message is expanded when
the rule is run.

R=binji@chromium.org

Review URL: https://codereview.chromium.org/18054027