aura: Make RenderWidgetHostViewAura::CopyFromCompositingSurface readback layer
Currently CopyFromCompositingSurface does an async texture readback of
current_surface_ which depends on the renderer compositor sending
textures to the browser compositor. Under ubercomp, however, the
current_surface_ will not exist.

The cc/ compositor now supports async readback of layers. This uses the
new mechanism to get the contents of the RWHVA's ui::Layer instead which
is always available. The result of this readback can be a texture or a
bitmap (when using software compositing). The result read by the
compositor will already be cropped.

If the result is a texture, we use the GLHelper to scale the result to
the desired size. If it is a bitmap, we use skia::ImageOperations::Resize
to do the same.

R=piman
BUG=162291

Review URL: https://chromiumcodereview.appspot.com/17971002