Make new dragged tab windows take on the restored window size.
Exploratory CL to cause new windows created as the result of a drag operation to be sized according to the restored window bounds. This applies to the "detach into new window while dragging" code path used by Aura, not the "make a semi-transparent preview window while dragging codepath" used on non-Aura Windows.

Also, always position tabs of such new windows to the left of the tabstrip. If this isn't done, then when dragging a tab from the right of a maximized window, the tab will be placed beyond the bounds of the restored window.

BUG=175448
TEST=Maximize a window in WinAura, then drag tabs off it.

Review URL: https://chromiumcodereview.appspot.com/17724003