Signin Page for OSX App Launcher.
Adds the SigninViewController under app_list. This controller loads the
signin view based on information provided by the delegate. The signin
view shows a blue button that opens a browser window to conduct the
signin flow. Also shown are some localized description text for signin,
and links: "Learn More" and "Settings".

If signin is required, the root applist view hides the regular contents
and adds the signin view as a subview. When signin is successful, the
subview is removed and the regular contents shown.

Screenshots at http://crbug.com/255237#4

BUG=138633,255237
TEST=Open the OSX App Launcher on a new profile, and test signin button,
and links to "Learn More" and "Settings" (which allows you to switch to
a different profile, e.g., one that is signed in). Also added
app_list_unittests for the root AppListViewControllerTest and the added
SigninViewController.

Review URL: https://chromiumcodereview.appspot.com/18326006