Correct RTL alignment of page title on chrome://policy
Currently the page header is aligned too far to the left, when viewing
the chrome://policy page in a right-to-left language. This looks
especially bad when scrolling down (see screenshots attached to bug).
This change aligns the page title to the right hand side in RTL mode.

BUG=253860

Review URL: https://chromiumcodereview.appspot.com/17587016