Hide autofill popup before WebContentsImpl is destroyed.
TabAutofillManagerDelegate was initiating the autofill popup hiding process in
its destructor. However, this process ends up calling back into the
WebContentsImpl instance via a virtual method call on the base WebContents
class. As TabAutofillManagerDelegate is a WebContentsUserData and thus has its
destructor called as part of the WebContents base class destructor, the call
into WebContents is occurring when the instance's vtable is that of
WebContents, where the method is pure virtual.

This CL changes the autofill popup hiding process to occur on receiving the
WebContentsDestroyed notification, while the vtable of the WebContents instance
is still that of WebContentsImpl.

BUG=254490
TEST=On Windows, cause an autofill popup to appear and close the tab. The app
should not crash.

Review URL: https://chromiumcodereview.appspot.com/18272015