[SPDY] Make SpdySessionPool keep track of available sessions
Split a SpdySessionPool's sessions into available ones and unavailable ones.
Keep all sessions in a set, and all the available sessions in a map keyed
by SpdySessionKey. This is in preparation for making SpdySessionPool
own a session even if it receives a GOAWAY frame.

Split SpdySessionPool::Remove() into two functions -- MakeSessionUnavailable()
and RemoveUnavailableSession(). For now, sessions call them at the same
time, but in the future RemoveUnavailableSession() may be called later
than MakeSessionUnavailable() (for the GOAWAY case).

Don't add a SpdySession to a pool if it encountered an error during
initialization. Also, don't do some work in SpdySession::InitializeWithSocket()
if initialization fails.

Rename functions and variables in SpdySessionPool to be concise and
consistent. Inline a bunch of one-off functions. Rewrite the loop
in SpdySessionPool::RemoveAliases().

BUG=255701

Review URL: https://chromiumcodereview.appspot.com/18600010