Pepper: Add FileRef messages for path information
Both FileIO and URLRequestInfo need to get path information for a FileRef resource (while in the renderer). This will introduce an extra round-trip to the browser in both paths.

It's possible that we'll do something more clever for FileIO (to go back to only one round trip), but I think we will need this for URLRequestInfo regardless.

BUG=225441

Review URL: https://chromiumcodereview.appspot.com/16975012