Add cache_dir to gclient recipes.
Also reduce concurrency for gclient sync in git mode. It appears that
too many simultaneous connections to chromium.googlesource.com will
cause weird 403 errors. Not sure if this is a local or a remote 
problem, but reduce concurrency for now to get things working.

BUG=

Review URL: https://chromiumcodereview.appspot.com/18660003