Fix a bug where the chrome_search sample extension attempted to search the old code search.
There are still issues with the suggestions onInputChanged, but it now can successfully perform a GET call on the address. The issue is that since it looks like the current Code Search serves up the results asynchronously, making grabbing any XML for meaningful information a challenge. I'll continue to put some time into this to see if there is a method of getting to that data.

BUG=111196

Review URL: https://chromiumcodereview.appspot.com/18147007