Support cancellation of pending media stream requests at UI side.
When a page has more than one media stream requests, they are currently stored in a queue
in MediaCaptureDevicesDispatcher till they are executed sequentially. This results in
annoying and confusing popup of infobar for each remaining request when browser navigates
away from the page or reload the page. 

This patch uses triplet (render_process_id, render_view_id, page_request_id) as identifier
for a request to allow the client to cancel the request.

BUG=252904
R=joaodasilva@chromium.org, joi@chromium.org, sky@chromium.org, xians@chromium.org

Review URL: https://codereview.chromium.org/18647011