cc: Make sure we check for completed tile uploads after activating pending tree.
The NotifyReadyToActivate signal produced by the tile manager
indicates that all tiles required for activation will be ready
to draw next time CheckForCompletedTileUploads() is called. The
pending tree can be activated once this signal is received but
we need to ensure that we always check for completed tiles
before we draw.

This makes LTHI::PrepareToDraw perform this check if not already
done by the scheduler.

BUG=257596

Review URL: https://chromiumcodereview.appspot.com/18854013