Register for invalidations when SyncManager exists
The SyncBackendHost forwards all invalidations directly to the
SyncManager.  This can cause problems if the SyncManager does not exist
at the time.

This CL tries to fix the problem by ensuring that the SyncBackendHost is
registered to receive invalidations only when the SyncManager actually
exists.  This makes it more difficult to know whether or not it is
necessary to unregister at shutdown, so this CL also adds a boolean flag
to track the SyncBackendHost's registration state.

BUG=256878

Review URL: https://chromiumcodereview.appspot.com/18551009