Rename RenderViewGone IPC/methods to better reflect reality.
The RenderViewGone IPC and methods are incorrectly named, as those are used when a renderer process has died. Since we are passing in termination status around and usually iterating over all Render{Widget|View}Host in a specific process, it actually applies to the whole process. This CL renames all applicable instances of this method and the IPC message associated with it.

BUG=258087

Review URL: https://chromiumcodereview.appspot.com/18339006