linux_aura: Implement SetOpacity() and fix menu host.
In r201447, I initially implemented SetOpacity(), assuming that X11's opacity was backward (subtract it from 255 instead of counting up from 0) because otherwise it broke display of menus, which were manually calling SetOpacity(0). This was reverted in r202910 once someone tried to rely on it and pointed out that it was wired backwards.

After some debugging, it was shown that win_aura never calls SetOpacity() on the menus. This brings the same behaviour to linux. This fixes the menus and relands the SetOpacity() change.

BUG=none

Review URL: https://chromiumcodereview.appspot.com/18355005