chromeos: Update power status code.
This updates ash's power-status code to deal with updated
PowerSupplyProperties sent by the power manager where:

a) the battery time-to-full estimate is set when the battery
   is charging and time-to-empty is set when the battery is
   discharging (previously, time-to-full was set whenever a
   charger was connected)

b) negative battery time estimates are used in situations
   where the current is close to zero, which would result in
   a huge estimate

It also updates ash to not display estimates above one day
and to not show an estimate when the battery is discharging
while a charger is connected.

Finally, it makes some minor improvements to battery-time
accessibility strings.

BUG=252948,254173,chrome-os-partner:20343

Review URL: https://chromiumcodereview.appspot.com/18324014