Fix JID refresh logic.
We should enable the JID refresh logic when the user connects to a host (since the entry might be stale), or when the user requests a reconnect (since the disconnection could be caused by the host going offline).

In either case, the connection attempt after the refresh should not result in a new refresh.

BUG=258188

Review URL: https://chromiumcodereview.appspot.com/18355008