add a chromium_builder_asan target
When I changed the ASAN builders to use 'all_webkit' in r202619 I made
them build a bunch of stuff they don't care about. This change adds
an asan builder target so that we can control which binaries are built
in the source tree. A follow-on change will switch the bots to use this.

R=inferno@chromium.org, jochen@chromium.org
BUG=258605

Review URL: https://chromiumcodereview.appspot.com/18066013