Cleanup swapped out RenderViewHosts when the process goes away.
Currently swapped out RVHs are not cleaned up for a given WebContents if the renderer hosting the associated RVs crashes. During webview postMessage, when sending a message from the embedder to the guest after recovering from a crash, we use the routing ID of the existing swapped out RVH which does not have a corresponding RV in the new renderer process. This produces a null e.source contentWindow and so the new guest has no way to communicate back to the embedder.

BUG=258993
Test=RenderViewHostManagerTest.CleanUpSwappedOutRVHOnProcessCrash

Review URL: https://chromiumcodereview.appspot.com/18512004