Pass BrowserContext to drive::MountPointProviderDelegate.
As a preparation to get rid of RemoteFileSystemProxy,
let MountPointProviderDelegate take BrowserContext instead of
ExternalMountPoints.
Currently, it is used to take its ExternalMountPoints, but will be used as a
key to take DriveIntegrationService, which manages the drive::FileSystem
instance.
The changes in drive::MountPointProviderDelegate will be done in following CLs.

BUG=253837
TEST=Ran unit_tests

Review URL: https://chromiumcodereview.appspot.com/18667002